+++ mod.ts
@@ -4,7 +4,6 @@
 
 /** property decorator */
 export { SerializeProperty } from "./serialize_property.ts";
-
 /** types */
 export type { TransformKey } from "./serializable.ts";
 /** abstract class */
@@ -16,7 +15,17 @@ export type {
   FromJSONStrategy,
   ToJSONStrategy,
 } from "./strategy/compose_strategy.ts";
-
 /** from json strategies */
 export { toSerializable } from "./strategy/from_json/to_serializable.ts";
 export { createDateStrategy, iso8601Date } from "./strategy/from_json/date.ts";
+/** polymorphic classes */
+export {
+  polymorphicClassFromJSON,
+  PolymorphicResolver,
+  PolymorphicSwitch,
+} from "./polymorphic/polymorphic.ts";
+/** types */
+export type {
+  InitializerFunction,
+  ResolverFunction,
+} from "./polymorphic/polymorphic.ts";