+++ CHANGELOG.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - new error_messages.ts file 
+- Added `PolymorphicResolver` and `PolymorphicSwitch` decorators, and 
+`polymorphicClassFromJSON` function for deserializing polymorphic data
+ structures
 
 ## [v0.4.0] - 2020-11-19